Restructured app, moved components to src folder, made that a module.
